Move trie cache to extra module
